requirement.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
docx2txt
python-dotenv
streamlit